Add map reset button
Because router.push doesn't do anything if the path passed to it is
the current path, in that scenario we need to reload the page in order
to get to the default zoom level with everything in bounds. This isn't
optimal, because of the use of 'window' which results in a full page
reload instead of a virtual DOM update.